+++ src/features/Conversation/Messages/Assistant/Tool/Render/CustomRender.tsx
@@ -1,4 +1,9 @@
-import { memo, useEffect } from 'react';
+import { ActionIcon } from '@lobehub/ui';
+import { App } from 'antd';
+import { Edit3Icon, PlayCircleIcon } from 'lucide-react';
+import { parse } from 'partial-json';
+import { memo, useCallback, useEffect, useState } from 'react';
+import { useTranslation } from 'react-i18next';
 import { Flexbox } from 'react-layout-kit';
 
 import PluginRender from '@/features/PluginsUI/Render';
@@ -7,6 +12,16 @@ import { chatSelectors } from '@/store/chat/selectors';
 import { ChatMessage } from '@/types/message';
 
 import Arguments from './Arguments';
+import KeyValueEditor from './KeyValueEditor';
+
+const safeParseJson = (str: string): Record<string, any> => {
+  try {
+    const obj = parse(str);
+    return typeof obj === 'object' && obj !== null ? obj : {};
+  } catch {
+    return {};
+  }
+};
 
 interface CustomRenderProps extends ChatMessage {
   requestArgs?: string;
@@ -25,7 +40,37 @@ const CustomRender = memo<CustomRenderProps>(
     setShowPluginRender,
     pluginError,
   }) => {
+    const { t } = useTranslation(['tool', 'common']);
     const [loading] = useChatStore((s) => [chatSelectors.isPluginApiInvoking(id)(s)]);
+    const [isEditing, setIsEditing] = useState(false);
+    const { message } = App.useApp();
+    const [updatePluginArguments, reInvokeToolMessage] = useChatStore((s) => [
+      s.updatePluginArguments,
+      s.reInvokeToolMessage,
+    ]);
+    const handleCancel = useCallback(() => {
+      setIsEditing(false);
+    }, []);
+
+    const handleFinish = useCallback(
+      async (editedObject: Record<string, any>) => {
+        if (!id) return;
+
+        try {
+          const newArgsString = JSON.stringify(editedObject, null, 2);
+
+          if (newArgsString !== requestArgs) {
+            await updatePluginArguments(id, editedObject, true);
+            await reInvokeToolMessage(id);
+          }
+          setIsEditing(false);
+        } catch (error) {
+          console.error('Error stringifying arguments:', error);
+          message.error(t('updateArgs.stringifyError'));
+        }
+      },
+      [requestArgs, id],
+    );
 
     useEffect(() => {
       if (!plugin?.type || loading) return;
@@ -49,8 +94,36 @@ const CustomRender = memo<CustomRenderProps>(
             pluginState={pluginState}
             type={plugin?.type}
           />
+        ) : isEditing ? (
+          <KeyValueEditor
+            initialValue={safeParseJson(requestArgs || '')}
+            onCancel={handleCancel}
+            onFinish={handleFinish}
+          />
         ) : (
-          <Arguments arguments={requestArgs} messageId={id} />
+          <Arguments
+            actions={
+              <>
+                <ActionIcon
+                  icon={Edit3Icon}
+                  onClick={() => {
+                    setIsEditing(true);
+                  }}
+                  size={'small'}
+                  title={t('edit', { ns: 'common' })}
+                />
+                <ActionIcon
+                  icon={PlayCircleIcon}
+                  onClick={async () => {
+                    await reInvokeToolMessage(id);
+                  }}
+                  size={'small'}
+                  title={t('run', { ns: 'common' })}
+                />
+              </>
+            }
+            arguments={requestArgs}
+          />
         )}
       </Flexbox>
     );
